backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
celery[redis]
pgvector
boto3
python-multipart
pytest
pytest-asyncio
pydantic[email]